refactor: name hook, inputs
