+++ .husky/pre-commit
@@ -1,10 +1,5 @@
 npm run format:fix
 npm run lint:fix
 
-if ! git diff --quiet
-  echo
-  exit 1
-fi
-
 npm run format
 npm run lint
